+++ src/clj/muuntaja/format/json.clj
@@ -4,21 +4,28 @@
   (:import (java.io InputStream
                     InputStreamReader
                     OutputStreamWriter
-                    OutputStream)))
+                    OutputStream)
+           (com.fasterxml.jackson.databind ObjectMapper)))
 
-(defn assert-options! [options]
-  (assert
-    (not (or (contains? options :key-fn) (contains? options :bigdecimals?)))
-    (str
-      "In Muuntaja 0.6.0+ the default JSON formatter has changed\n"
-      "from Cheshire to Jsonita. Changed options:\n\n"
-      "  :key-fn       => :encode-key-fn & :decode-key-fn\n"
-      "  :bigdecimals? => :bigdecimals\n"
-      options "\n")))
+(defn object-mapper! [{:keys [mapper] :as options}]
+  (cond
+    (instance? ObjectMapper mapper)
+    mapper
+
+    (or (contains? options :key-fn) (contains? options :bigdecimals?))
+    (throw (AssertionError.
+             (str
+               "In Muuntaja 0.6.0+ the default JSON formatter has changed\n"
+               "from Cheshire to Jsonita. Changed options:\n\n"
+               "  :key-fn       => :encode-key-fn & :decode-key-fn\n"
+               "  :bigdecimals? => :bigdecimals\n"
+               options "\n")))
+
+    :else
+    (j/object-mapper (dissoc options :mapper))))
 
 (defn make-json-decoder [options]
-  (assert-options! options)
-  (let [mapper (j/object-mapper options)]
+  (let [mapper (object-mapper! options)]
     (fn [x ^String charset]
       (if (string? x)
         (j/read-value x mapper)
@@ -27,17 +34,14 @@
           (j/read-value (InputStreamReader. ^InputStream x charset) mapper))))))
 
 (defn make-json-encoder [options]
-  (assert-options! options)
-  (let [mapper (j/object-mapper options)]
+  (let [mapper (object-mapper! options)]
     (fn [data ^String charset]
-      (protocols/->ByteResponse
-        (if (.equals "utf-8" charset)
-          (j/write-value-as-bytes data mapper)
-          (.getBytes ^String (j/write-value-as-string data mapper) charset))))))
+      (if (.equals "utf-8" charset)
+        (j/write-value-as-bytes data mapper)
+        (.getBytes ^String (j/write-value-as-string data mapper) charset)))))
 
 (defn make-streaming-json-encoder [options]
-  (assert-options! options)
-  (let [mapper (j/object-mapper options)]
+  (let [mapper (object-mapper! options)]
     (fn [data ^String charset]
       (protocols/->StreamableResponse
         (fn [^OutputStream output-stream]