+++ src/clj/muuntaja/protocols.clj
@@ -2,7 +2,8 @@
   (:require [clojure.java.io :as io]
             [muuntaja.util :as util])
   (:import (clojure.lang IFn AFn)
-           (java.io ByteArrayOutputStream ByteArrayInputStream InputStreamReader BufferedReader InputStream Writer OutputStream)))
+           (java.io ByteArrayOutputStream ByteArrayInputStream InputStreamReader BufferedReader InputStream Writer OutputStream FileInputStream File)
+           (java.nio ByteBuffer)))
 
 (deftype ByteResponse [bytes])
 
@@ -17,6 +18,10 @@
 (util/when-ns
   'ring.core.protocols
   (extend-protocol ring.core.protocols/StreamableResponseBody
+    (Class/forName "[B")
+    (write-body-to-stream [this _ output-stream]
+      (.write ^OutputStream output-stream ^bytes this))
+
     ByteResponse
     (write-body-to-stream [this _ output-stream]
       (.write ^OutputStream output-stream ^bytes (.bytes this)))
@@ -69,6 +74,12 @@
   (-input-stream ^java.io.InputStream [this]))
 
 (extend-protocol IntoInputStream
+  (Class/forName "[B")
+  (-input-stream [this] (ByteArrayInputStream. this))
+
+  File
+  (-input-stream [this] (FileInputStream. this))
+
   InputStream
   (-input-stream [this] this)