+++ test/muuntaja/core_test.clj
@@ -2,15 +2,18 @@
   (:require [clojure.test :refer :all]
             [muuntaja.core :as m]
             [clojure.string :as str]
-            [muuntaja.protocols :as protocols]
             [muuntaja.format.cheshire :as cheshire-format]
             [muuntaja.format.msgpack :as msgpack-format]
             [muuntaja.format.yaml :as yaml-format]
-            [jsonista.core :as j])
+            [jsonista.core :as j]
+            [clojure.java.io :as io]
+            [muuntaja.protocols :as protocols]
+            [muuntaja.util :as util])
   (:import (java.nio.charset Charset)
-           (java.io ByteArrayInputStream)))
+           (java.io FileInputStream)
+           (java.nio.file Files)))
 
-(defn- to-byte-stream [x charset] (ByteArrayInputStream. (.getBytes x charset)))
+(defn- to-byte-stream [x charset] (util/byte-stream (.getBytes x charset)))
 
 (defn set-jvm-default-charset! [charset]
   (System/setProperty "file.encoding" charset)
@@ -77,7 +80,7 @@
         (is (= "UTF-16" (str (Charset/defaultCharset)))))))
 
   (testing "on empty input"
-    (let [empty (fn [] (ByteArrayInputStream. (byte-array 0)))
+    (let [empty (fn [] (util/byte-stream (byte-array 0)))
           m2 (m/create
                (-> m/default-options
                    (msgpack-format/with-msgpack-format)
@@ -194,7 +197,7 @@
   (testing "adding new format"
     (let [format "application/upper"
           upper-case-format {:decoder (fn [s _] (str/lower-case (slurp s)))
-                             :encoder (fn [s _] (protocols/-input-stream (str/upper-case s)))}
+                             :encoder (fn [s _] (.getBytes (str/upper-case s)))}
           m (m/create
               (-> m/default-options
                   (assoc-in [:formats format] upper-case-format)))
@@ -227,9 +230,25 @@
                             m/default-options
                             [:formats "application/json" :decoder-opts]
                             {:decode-key-fn false}))
-                        "application/json")]
+                        "application/json")
+          decode-json2 (m/decoder
+                         (m/create
+                           (assoc-in
+                             m/default-options
+                             [:formats "application/json" :opts]
+                             {:decode-key-fn false}))
+                         "application/json")
+          decode-json3 (m/decoder
+                         (m/create
+                           (assoc-in
+                             m/default-options
+                             [:formats "application/json" :opts]
+                             {:mapper (j/object-mapper {:decode-key-fn false})}))
+                         "application/json")]
       (is (= {:kikka true} (decode-json-kw "{\"kikka\":true}")))
-      (is (= {"kikka" true} (decode-json "{\"kikka\":true}")))))
+      (is (= {"kikka" true} (decode-json "{\"kikka\":true}")))
+      (is (= {"kikka" true} (decode-json2 "{\"kikka\":true}")))
+      (is (= {"kikka" true} (decode-json3 "{\"kikka\":true}")))))
 
   (testing "overriding invalid adapter options fails"
     (is (thrown?
@@ -267,3 +286,21 @@
                   (assoc-in
                     [:formats "application/json" :decoder-opts]
                     {:key-fn false}))))))))
+
+(deftest slurp-test
+  (let [file (io/file "dev-resources/json10b.json")
+        expected (slurp file)]
+    (testing "bytes"
+      (is (= expected (m/slurp (Files/readAllBytes (.toPath file))))))
+    (testing "File"
+      (is (= expected (m/slurp file))))
+    (testing "InputStream"
+      (is (= expected (m/slurp (FileInputStream. file)))))
+    (testing "StreamableResponse"
+      (is (= expected (m/slurp (protocols/->StreamableResponse (partial io/copy file))))))
+    (testing "ByteReponse"
+      (is (= expected (m/slurp (protocols/->ByteResponse (Files/readAllBytes (.toPath file)))))))
+    (testing "String"
+      (is (= expected (m/slurp expected))))
+    (testing "nil"
+      (is (= nil (m/slurp nil))))))